+++ adapter/spark/src/main/java/com/asml/apa/wta/spark/listener/TaskStageBaseListener.java
@@ -7,7 +7,6 @@ import java.util.concurrent.ConcurrentHashMap;
 import lombok.Getter;
 import org.apache.spark.SparkContext;
 import org.apache.spark.scheduler.SparkListenerJobStart;
-import org.apache.spark.scheduler.SparkListenerStageCompleted;
 
 /**
  * This abstract class is a base class for the task and stage level listeners.
@@ -19,7 +18,7 @@ import org.apache.spark.scheduler.SparkListenerStageCompleted;
 public abstract class TaskStageBaseListener extends AbstractListener<Task> {
 
   @Getter
-  protected final Map<Integer, Integer> stageIdsToJobs = new ConcurrentHashMap<>();
+  protected final Map<Long, Long> stageToJob = new ConcurrentHashMap<>();
 
   /**
    * Constructor for the stage-level listener.
@@ -34,8 +33,7 @@ public abstract class TaskStageBaseListener extends AbstractListener<Task> {
   }
 
   /**
-   * This method is called every time a job starts.
-   * In the context of the WTA, this is a workflow.
+   * This method is called every time a job starts. In the context of the WTA, this is a workflow.
    *
    * @param jobStart The object corresponding to information on job start.
    * @author Henry Page
@@ -43,20 +41,7 @@ public abstract class TaskStageBaseListener extends AbstractListener<Task> {
    */
   @Override
   public void onJobStart(SparkListenerJobStart jobStart) {
-    // stage ids are always unique
-    jobStart.stageInfos().foreach(stageInfo -> stageIdsToJobs.put(stageInfo.stageId() + 1, jobStart.jobId() + 1));
-  }
-
-  /**
-   * Callback for when a stage ends.
-   *
-   * @param stageCompleted The stage completion event
-   * @author Henry Page
-   * @since 1.0.0
-   */
-  @Override
-  public void onStageCompleted(SparkListenerStageCompleted stageCompleted) {
-    // all tasks are guaranteed to be completed, so we can remove the stage id to reduce memory usage.
-    stageIdsToJobs.remove(stageCompleted.stageInfo().stageId() + 1);
+    long jobId = jobStart.jobId() + 1;
+    jobStart.stageInfos().foreach(stageInfo -> stageToJob.put((long) stageInfo.stageId() + 1, jobId));
   }
 }