+++ adapter/spark/src/test/java/com/asml/apa/wta/spark/listener/StageLevelListenerIntegrationTest.java
@@ -6,21 +6,25 @@ import com.asml.apa.wta.spark.BaseSparkJobIntegrationTest;
 import org.junit.jupiter.api.Test;
 
 public class StageLevelListenerIntegrationTest extends BaseSparkJobIntegrationTest {
+
   @Test
   void testGetStageMetricsHasTasksAfterSparkJobAndYieldsNoErrors() {
-    sut.registerStageListener();
+    sut1.registerStageListener();
     invokeJob();
-    assertThat(sut.getStageLevelListener().getProcessedObjects()).isNotEmpty();
+    invokeJob();
+    stopJob();
+    assertThat(sut1.getStageLevelListener().getProcessedObjects()).isNotEmpty();
   }
 
   @Test
-  void runningAJobShouldClearTheMapOfEntriesAfterStageIsDoneWorkflowIdShouldBeInitialised() {
-    sut.registerStageListener();
+  void runningAJobShouldInitializeWorkflow() {
+    sut1.registerStageListener();
     invokeJob();
-    assertThat(((StageLevelListener) sut.getStageLevelListener()).getStageIdsToJobs())
-        .isEmpty();
-    assertThat(sut.getStageLevelListener().getProcessedObjects()).hasSizeGreaterThanOrEqualTo(1);
-    assertThat(sut.getStageLevelListener().getProcessedObjects().get(0).getWorkflowId())
+    invokeJob();
+    stopJob();
+    assertThat(sut1.getStageLevelListener().getStageToJob()).isNotEmpty();
+    assertThat(sut1.getStageLevelListener().getProcessedObjects()).hasSizeGreaterThanOrEqualTo(1);
+    assertThat(sut1.getStageLevelListener().getProcessedObjects().get(0).getWorkflowId())
         .isInstanceOf(Long.class)
         .isNotNull();
   }