+++ connector/hdfs/src/test/java/com/asml/apa/wta/hdfs/io/OutputFileFactoryWithHdfsTest.java
@@ -5,10 +5,13 @@ import static org.assertj.core.api.Assertions.assertThat;
 import com.asml.apa.wta.core.io.OutputFile;
 import com.asml.apa.wta.core.io.OutputFileFactory;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.api.condition.EnabledOnOs;
+import org.junit.jupiter.api.condition.OS;
 
 class OutputFileFactoryWithHdfsTest {
 
   @Test
+  @EnabledOnOs({OS.WINDOWS})
   void createSimpleLocation() {
     OutputFileFactory factory = new OutputFileFactory();
     OutputFile file = factory.create("C:/path/to/output");
@@ -16,6 +19,7 @@ class OutputFileFactoryWithHdfsTest {
   }
 
   @Test
+  @EnabledOnOs({OS.LINUX, OS.MAC})
   void creatRelativeLocation() {
     OutputFileFactory factory = new OutputFileFactory();
     OutputFile file = factory.create("path/to/output");